Add files via upload
Bioinformatics algorithm called Edit Distance. Used to determine how much two strings (or genomes) are different from one another. The implementation takes in two strings and returns the edit distance, the printed graph used to determine edit distance, and the global alignment of the strings. The implementation used a dynamic 2d list to determine the edit distance and iterated from the end of the graph to the start to print out the global alignment.