Add files via upload
This is a group project I did with other students at my school using the scrum process. Essentially this was made over 4 iterations with a group of 5 students over the school semester. We met 3-4 times a week and discussed our process whether people needed help or to decide goals for the next iteration. The project itself was creating a fully working chess gain with a simple ai. It was created in python using Pygame, and uses simple animations and UI design. The ai is based off the alpha beta pruning algorithm and searches to a maximum depth of 3.